feat:comment 댓글 작성자 검증 로직 추가 (#37)
* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* feat:project 유효성 로직 추가  #12

* feat:comment 댓글 기본 CRUD #18

* style:불필요한 파일 제거

* feat:comment 댓글 CRUD 기능 구현 #18

* feat:comment 프로젝트 end_date 순 정렬, 검색 기능 구현 #12

* feat:comment 프로젝트 end_date 순 정렬, 검색 기능 구현 #12

* feat:comment 댓글 crud 로직 수정

* feat:comment crud 수정

* feat:comment 댓글 멤버 검증 로직 구현

+++ src/main/java/com/douzone/prosync/comment/controller/CommentController.java
@@ -7,6 +7,8 @@ import com.douzone.prosync.comment.dto.response.GetCommentsResponse;
 import com.douzone.prosync.comment.entity.Comment;
 import com.douzone.prosync.comment.service.CommentService;
 import com.douzone.prosync.common.PageResponseDto;
+import com.douzone.prosync.exception.ApplicationException;
+import com.douzone.prosync.exception.ErrorCode;
 import io.swagger.annotations.*;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -74,7 +76,13 @@ public class CommentController {
             @RequestBody @Valid CommentPatchDto dto,
             @Parameter(hidden = true) @ApiIgnore Principal principal) {
 
-//        commentService.checkMember(Long.valueOf(principal.getName()));
+        // 작성한 멤버인지 검증
+        boolean isCommentOwner = commentService.checkMember(commentId,Long.valueOf(principal.getName()));
+
+        if(!isCommentOwner) {
+            throw new ApplicationException(ErrorCode.ACCESS_FORBIDDEN);
+        }
+
         dto.setCommentId(commentId);
         commentService.update(dto);
         return new ResponseEntity(new CommentSimpleResponse(commentId), HttpStatus.OK);
@@ -119,6 +127,14 @@ public class CommentController {
     public ResponseEntity deleteComment(
             @Parameter(description = "댓글 식별자", required = true, example = "1") @PathVariable("comment-id") Integer commentId,
             @ApiIgnore Principal principal) {
+
+        // 작성한 멤버인지 검증
+        boolean isCommentOwner = commentService.checkMember(commentId,Long.valueOf(principal.getName()));
+
+        if(!isCommentOwner) {
+            throw new ApplicationException(ErrorCode.ACCESS_FORBIDDEN);
+        }
+
         commentService.delete(commentId);
         return new ResponseEntity(new CommentSimpleResponse(commentId), HttpStatus.NO_CONTENT);
     }
